Content-moderator-image-1.3/backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
pymupdf
ollama
fastapi
uvicorn
faiss-cpu
ollama
pymupdf
numpy
sentence-transformers
streamlit
requests
python-multipart